REFACTOR: refactor tests to use TestSetup module
